Update the General Annual Grant Profiling pages
Pages: http://localhost:3000/v1/funding-management/gag/provider-funding-overview-gag#profiling
·      The 3 profiling pages (linked to from the link above) require the following updates

·      In the Summary List at the top of the page:

a.     Make ‘Total allocation’ a negative value

                                               i.     E.g. £-£2,600,000

b.     Make ‘Balance available for profiling’ a negative value

·      Remove the Paid tags

·      In the ‘Value’ column:

a.     Set all values to ‘£0’

b.     Except for August which will have a value that = 100% of Total allocation

                                               i.     E.g. -£2,600,000

·      Update ‘To be carried forward’ to read ‘Amount to be carried forward’

a.     For this row, also add in a value of ‘10%’ into the ‘Per cent’ column

b.     Amount to be carried forward ‘value’ set to £0

·      Change total allocation to ‘Total profiled instalments’

a.     Update value to be ‘-£2,600,000’

·      Add a last row, called ‘Total’

a.     This should sum up all rows including ‘amount to be carried forward’ and Total profiled instalments’ added together

·      Make ‘Amount to be carried forward’ and ‘Total profiled instalments’ items in the ‘values’ column to be bold

·      Remove the ‘Change to rule based profile’ button

·      Remove the ‘export to csv’ link

The changes listed in have been applied to the ‘edit’ pages