Calculation script and SQL data
I've removed the full screen includes file and all the expand/minimise icons on the calculation script screen as this isn't compatible with the Monaco Editor. Inside the SQL Data for the specifications that are not chosen for funding, spec 2,3 I've removed the details component and the action for the current and latest release state allocations, replacing it inset text saying these aren't chosen for funding.